requirements.txt
flask
gunicorn
numpy
pandas
scikit-learn
matplotlib
seaborn
wordcloud
Sastrawi
tensorflow-cpu==2.21.0